Move the model package one level up (#290)
