+++ model/common.go
