+++ model/stacktrace_frame.go
