+++ model/stacktrace_frame_test.go
