+++ processor/error/event_test.go
@@ -10,7 +10,7 @@ import (
 
 	"time"
 
-	m "github.com/elastic/apm-server/processor/model"
+	m "github.com/elastic/apm-server/model"
 	"github.com/elastic/beats/libbeat/common"
 )
