+++ processor/error/payload.go
@@ -1,8 +1,8 @@
 package error
 
 import (
+	m "github.com/elastic/apm-server/model"
 	pr "github.com/elastic/apm-server/processor"
-	m "github.com/elastic/apm-server/processor/model"
 	"github.com/elastic/beats/libbeat/beat"
 	"github.com/elastic/beats/libbeat/logp"
 	"github.com/elastic/beats/libbeat/monitoring"
