+++ processor/transaction/payload_test.go
@@ -8,7 +8,7 @@ import (
 
 	"time"
 
-	m "github.com/elastic/apm-server/processor/model"
+	m "github.com/elastic/apm-server/model"
 	"github.com/elastic/beats/libbeat/common"
 )
